+++ web-frontend/modules/database/components/view/grid/GridViewSection.vue
@@ -4,10 +4,12 @@
       <GridViewHead
         :table="table"
         :view="view"
-        :fields="visibleFields"
+        :fields="fields"
         :include-field-width-handles="includeFieldWidthHandles"
         :include-row-details="includeRowDetails"
         :include-add-field="includeAddField"
+        :read-only="readOnly"
+        :store-prefix="storePrefix"
         @refresh="$emit('refresh', $event)"
         @dragging="
           canOrderFields &&
@@ -17,19 +19,24 @@
       <div ref="body" class="grid-view__body">
         <div class="grid-view__body-inner">
           <GridViewPlaceholder
-            :fields="visibleFields"
+            :fields="fields"
             :include-row-details="includeRowDetails"
+            :store-prefix="storePrefix"
           ></GridViewPlaceholder>
           <GridViewRows
             :table="table"
             :view="view"
-            :fields="visibleFields"
+            :fields="fields"
             :include-row-details="includeRowDetails"
+            :read-only="readOnly"
+            :store-prefix="storePrefix"
             v-on="$listeners"
           ></GridViewRows>
           <GridViewRowAdd
-            :fields="visibleFields"
+            v-if="!readOnly"
+            :fields="fields"
             :include-row-details="includeRowDetails"
+            :store-prefix="storePrefix"
             v-on="$listeners"
           ></GridViewRowAdd>
         </div>
@@ -41,8 +48,9 @@
     <GridViewFieldDragging
       ref="fieldDragging"
       :view="view"
-      :fields="visibleFields"
+      :fields="fields"
       :container-width="width"
+      :store-prefix="storePrefix"
       @scroll="$emit('scroll', $event)"
     ></GridViewFieldDragging>
   </div>
@@ -55,7 +63,6 @@ import GridViewRows from '@baserow/modules/database/components/view/grid/GridVie
 import GridViewRowAdd from '@baserow/modules/database/components/view/grid/GridViewRowAdd'
 import GridViewFieldDragging from '@baserow/modules/database/components/view/grid/GridViewFieldDragging'
 import gridViewHelpers from '@baserow/modules/database/mixins/gridViewHelpers'
-import { GridViewType } from '@baserow/modules/database/viewTypes'
 
 export default {
   name: 'GridViewSection',
@@ -100,51 +107,18 @@ export default {
       required: false,
       default: () => false,
     },
+    readOnly: {
+      type: Boolean,
+      required: true,
+    },
   },
   computed: {
-    /**
-     * Returns only the visible fields in the correct order.
-     */
-    visibleFields() {
-      return this.fields
-        .filter((field) => {
-          const exists = Object.prototype.hasOwnProperty.call(
-            this.fieldOptions,
-            field.id
-          )
-          return !exists || (exists && !this.fieldOptions[field.id].hidden)
-        })
-        .sort((a, b) => {
-          const orderA = this.fieldOptions[a.id]
-            ? this.fieldOptions[a.id].order
-            : GridViewType.getMaxPossibleOrderValue()
-          const orderB = this.fieldOptions[b.id]
-            ? this.fieldOptions[b.id].order
-            : GridViewType.getMaxPossibleOrderValue()
-
-          // First by order.
-          if (orderA > orderB) {
-            return 1
-          } else if (orderA < orderB) {
-            return -1
-          }
-
-          // Then by id.
-          if (a.id < b.id) {
-            return -1
-          } else if (a.id > b.id) {
-            return 1
-          } else {
-            return 0
-          }
-        })
-    },
     /**
      * Calculates the total width of the whole section based on the fields and the
      * given options.
      */
     width() {
-      let width = Object.values(this.visibleFields).reduce(
+      let width = Object.values(this.fields).reduce(
         (value, field) => this.getFieldWidth(field.id) + value,
         0
       )