+++ changelog.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## Released (2021-04-08)
 
 * Added support for importing tables from XML files.
 * Added support for different character encodings when importing CSV files.
@@ -25,6 +25,7 @@
 * Add Phone Number field.
 * Add support for Date, Number and Single Select fields to the Contains and Not Contains view 
   filters.
+* Searching all rows can now be done by clicking the new search icon in the top right.
 
 ## Released (2021-03-01)